implementerat fireplace-metoder till forest-klassen
