+++ Adventure_Game/Cave.java
@@ -1,0 +1,8 @@
+package Adventure_Game;
+
+public class Cave {
+    public Cave(){
+        
+    }
+    
+}
