+++ Adventure_Game/City.java
@@ -1,0 +1,8 @@
+package Adventure_Game;
+
+public class City{
+    public City(){
+        
+    }
+    
+}
