+++ Adventure_Game/Hospital.java
@@ -1,0 +1,8 @@
+package Adventure_Game;
+
+public class Hospital {
+    public Hospital(){
+       
+    }
+    
+}
