+++ Adventure_Game/Mall.java
@@ -1,0 +1,8 @@
+package Adventure_Game;
+
+public class Mall {
+    public Mall(){
+        
+    }
+    
+}
