+++ Adventure_Game/Policestation.java
@@ -1,0 +1,8 @@
+package Adventure_Game;
+
+public class Policestation {
+    public Policestation(){
+        
+    }
+    
+}
