--- Cave.java
@@ -1,8 +1,0 @@
-package AdventureGame;
-
-public class Cave extends Environment {
-    public Cave(){
-        description = "";
-    }
-    
-}
