--- City.java
@@ -1,8 +1,0 @@
-package AdventureGame;
-
-public class City extends Environment{
-    public City(){
-        description ="";
-    }
-    
-}
