--- Environment.java
@@ -1,6 +1,0 @@
-package AdventureGame;
-
-public abstract class Environment {
-    protected String description;
-    
-}
