--- Hospital.java
@@ -1,8 +1,0 @@
-package AdventureGame;
-
-public class Hospital extends Environment{
-    public Hospital(){
-        description = "";
-    }
-    
-}
