--- Mall.java
@@ -1,8 +1,0 @@
-package AdventureGame;
-
-public class Mall extends Environment {
-    public Mall(){
-        description = "";
-    }
-    
-}
