--- Policestation.java
@@ -1,8 +1,0 @@
-package AdventureGame;
-
-public class Policestation extends Environment{
-    public Policestation(){
-        description = "";
-    }
-    
-}
